@@ -2,73 +2,73 @@
     "name": "demo",
     "version": "0.0.0",
     "license": "MIT",
+
     "scripts": {
         "ng": "ng",
         "dev": "ng serve --port 4201",
-        "test": "ng test",
+        "test-lib": "ng test natural",
+        "test-demo": "ng test demo",
         "lint": "ng lint",
         "e2e": "ng e2e",
         "build-lib": "ng build natural && cp -v README.md dist/natural/ && scss-bundle -e ./projects/natural/src/lib/_natural.theme.scss -d ./dist/natural/theming/_natural.theme.scss && cd dist/natural && yarn version --no-git-tag-version --new-version `git describe --tags --always`",
-        "build-docs": "yarn build-lib && ng build demo --prod --aot --base-href /natural/ && cp dist/demo/index.html dist/demo/404.html",
+        "build-demo": "yarn build-lib && ng build demo --prod --aot --base-href /natural/ && cp dist/demo/index.html dist/demo/404.html",
         "llink": "cd dist/natural && (yarn unlink ; yarn link)"
     },
     "private": true,
     "dependencies": {
-        "@angular/animations": "~7.2.0",
-        "@angular/cdk": "^7.3.7",
-        "@angular/common": "~7.2.0",
-        "@angular/compiler": "~7.2.0",
-        "@angular/core": "~7.2.0",
-        "@angular/flex-layout": "^7.0.0-beta.24",
-        "@angular/forms": "~7.2.0",
-        "@angular/material": "^7.3.7",
-        "@angular/platform-browser": "~7.2.0",
-        "@angular/platform-browser-dynamic": "~7.2.0",
-        "@angular/router": "~7.2.0",
+        "@angular/animations": "~8.2.0",
+        "@angular/cdk": "^8.1.2",
+        "@angular/common": "~8.2.0",
+        "@angular/compiler": "~8.2.0",
+        "@angular/core": "~8.2.0",
+        "@angular/flex-layout": "^8.0.0-beta.26",
+        "@angular/forms": "~8.2.0",
+        "@angular/material": "^8.1.2",
+        "@angular/platform-browser": "~8.2.0",
+        "@angular/platform-browser-dynamic": "~8.2.0",
+        "@angular/router": "~8.2.0",
         "apollo": "^2.8.3",
-        "apollo-angular": "^1.5.0",
-        "apollo-angular-link-http": "^1.6.0",
-        "apollo-angular-link-http-batch": "^1.5.0",
+        "apollo-angular": "^1.6.0",
+        "apollo-angular-link-http": "^1.8.0",
+        "apollo-angular-link-http-batch": "^1.7.0",
         "apollo-cache-inmemory": "^1.5.1",
         "apollo-client": "^2.5.1",
         "apollo-link": "^1.2.11",
         "apollo-link-error": "^1.1.10",
         "apollo-link-schema": "^1.2.2",
         "apollo-upload-client": "^10.0.0",
-        "core-js": "^2.5.4",
         "graphql": "^14.2.1",
         "graphql-tag": "^2.10.1",
         "graphql-tools": "^4.0.4",
         "lodash": "^4.17.13",
-        "rxjs": "~6.3.3",
-        "tslib": "^1.9.0",
-        "zone.js": "~0.8.26"
+        "rxjs": "~6.5.2",
+        "tslib": "^1.10.0",
+        "zone.js": "~0.9.1"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "~0.13.0",
-        "@angular-devkit/build-ng-packagr": "~0.13.0",
-        "@angular/cli": "~7.3.7",
-        "@angular/compiler-cli": "~7.2.0",
-        "@angular/language-service": "~7.2.0",
-        "@types/jasmine": "~2.8.8",
-        "@types/jasminewd2": "~2.0.3",
+        "@angular-devkit/build-angular": "~0.802.0",
+        "@angular-devkit/build-ng-packagr": "~0.802.0",
+        "@angular/cli": "~8.2.0",
+        "@angular/compiler-cli": "~8.2.0",
+        "@angular/language-service": "~8.2.0",
         "@types/node": "~8.9.4",
-        "codelyzer": "~4.5.0",
-        "jasmine-core": "~2.99.1",
+        "@types/jasmine": "~3.3.8",
+        "@types/jasminewd2": "~2.0.3",
+        "codelyzer": "^5.0.0",
+        "jasmine-core": "~3.4.0",
         "jasmine-spec-reporter": "~4.2.1",
-        "karma": "~4.0.0",
+        "karma": "~4.1.0",
         "karma-chrome-launcher": "~2.2.0",
         "karma-coverage-istanbul-reporter": "~2.0.1",
-        "karma-jasmine": "~1.1.2",
-        "karma-jasmine-html-reporter": "^0.2.2",
-        "ng-packagr": "^4.2.0",
+        "karma-jasmine": "~2.0.1",
+        "karma-jasmine-html-reporter": "^1.4.0",
+        "ng-packagr": "^5.3.0",
         "protractor": "~5.4.0",
-        "puppeteer": "^1.14.0",
+        "puppeteer": "^1.19.0",
         "scss-bundle": "^2.5.1",
         "ts-node": "~7.0.0",
-        "tsickle": ">=0.34.0",
-        "tslib": "^1.9.0",
-        "tslint": "~5.11.0",
-        "typescript": "~3.2.2"
+        "tsickle": "^0.36.0",
+        "tslint": "~5.15.0",
+        "typescript": "~3.5.3"
     }
 }
