@@ -4,9 +4,9 @@
   "license": "MIT",
   "repository": "github:Ecodev/natural",
   "peerDependencies": {
-    "@angular/common": "^7.2.0",
-    "@angular/core": "^7.2.0",
-    "@angular/material": "^7.3.0",
+    "@angular/common": "^8.2.0",
+    "@angular/core": "^8.2.0",
+    "@angular/material": "^8.1.0",
     "apollo": "^2.6.2",
     "apollo-angular": "^1.5.0",
     "apollo-angular-link-http": "^1.4.0",
