@@ -19,7 +19,7 @@ module.exports = function (config) {
             clearContext: false, // leave Jasmine Spec Runner output visible in browser
         },
         coverageIstanbulReporter: {
-            dir: require('path').join(__dirname, '../../coverage/natural'),
+            dir: require('path').join(__dirname, '../coverage/natural'),
             reports: ['html', 'lcovonly', 'text-summary'],
             fixWebpackSourcePaths: true,
         },
